Create a family event object

Create a family event object

Usage

FamilyEvent(
  fact_type = character(0),
  fact_val = character(0),
  fact_desc = character(0),
  date = character(0),
  place = NULL,
  address = NULL,
  phone_numbers = character(0),
  emails = character(0),
  faxes = character(0),
  web_pages = character(0),
  agency = character(0),
  relig_affil = character(0),
  cause = character(0),
  confidential = FALSE,
  locked = FALSE,
  private = FALSE,
  date_sort = character(0),
  associations = list(),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  media_links = list(),
  unique_ids = character(0),
  husb_age = character(0),
  husb_age_phrase = character(0),
  wife_age = character(0),
  wife_age_phrase = character(0)
)

Arguments

fact_type

A code indicating the type of fact. This must be taken from one of val_individual_event_types(), val_individual_attribute_types(), val_family_event_types(), or val_family_attribute_types(). A generic event ("EVEN") or attribute ("FACT") can also be defined for more bespoke facts. See fact_rules_df() for the set of rules surrounding the codes allowed.

fact_val

A value associated with the fact. For example for "NCHI" this would be the number of children. See fact_rules_df() for the set of rules surrounding the need for values and the values allowed.

fact_desc

A further classification of the fact. This is required for generic events or attributes. See fact_rules_df() for the set of rules surrounding the need for this.

date

The date given either as a formatted GEDCOM string, or a DateValue() object.

place

The associated place. This can either be a Place() object or a character string (a comma-separated string of region names, ordered from smallest to largest).

address

The address given either as a Address() object or as a character string. This would be as written on a mailing label with new lines separated by \n.

phone_numbers

A character vector of phone numbers.

emails

A character vector of email addresses.

faxes

A character vector of fax numbers.

web_pages

A character vector of web page URLs.

agency

The organization, institution, corporation, person, or other entity that has responsibility for the associated fact. Examples are an employer of a person of an associated occupation, or an educational establishment that has awarded a scholastic award.

relig_affil

A religious denomination associated with the fact.

cause

Used in special cases to record the reasons which precipitated the fact (e.g. cause of death).

confidential

A logical value indicating whether the associated record/fact should be treated as confidential. This allows them to be excluded on export.

locked

A logical value indicating whether the associated record/fact should be treated as read-only.

private

A logical value indicating whether the associated record/fact should be treated as private. This allows them to be excluded on export.

date_sort

The date given either as a formatted GEDCOM string, or a DateSorting() object.

associations

Associated individuals. This can either be a Association() object or a list of them.

note_xrefs

A character vector of relevant note record cross-reference identifiers.

notes

Associated notes. This can either be a Note() object, a list of them, or a character vector of notes.

citations

Associated sources. This can either be a SourceCitation() object, a list of them, or a character vector of XREFs of source records.

media_links

Associated multimedia. This can either be a MediaLink() object, a list of them, or a character vector of XREFs of multimedia records.

unique_ids

A character vector of enduring and globally-unique identifiers. These need to be formatted in line with RFC 4122 and can be generated with uuid::UUIDgenerate().

husb_age

A character string that indicates the age in years, months, weeks and/or days that the husband was at the time of the fact. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 1w 3d". Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age in the corresponding phrase parameter.

husb_age_phrase

Free text information that cannot be expressed in the husband's age.

wife_age

A character string that indicates the age in years, months, weeks and/or days that the wife was at the time of the fact. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 1w 3d". Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age in the corresponding phrase parameter.

wife_age_phrase

Free text information that cannot be expressed in the wife's age.

Value

An S7 object representing a GEDCOM FAMILY_EVENT_STRUCTURE.