Create a GEDCOM object

You shouldn't need to use this directly to create new GEDCOM objects. Instead, use new_gedcom() which populates relevant defaults.

Usage

GedcomS7(
  gedcom_version = character(0),
  ext_tags = character(0),
  source = NULL,
  destination = character(0),
  creation_date = character(0),
  creation_time = character(0),
  subm_xref = character(0),
  gedcom_copyright = character(0),
  default_language = character(0),
  default_place_form = character(0),
  notes = list(),
  note_xrefs = character(0),
  update_change_dates = FALSE,
  add_creation_dates = FALSE,
  records = GEDCOMRecords()
)

Arguments

gedcom_version

The version number of the official specification that this GEDCOM conforms to. This must include the major and minor version (for example, “7.0”); it may include the patch as well (for example, “7.0.1”), but doing so is not required.

ext_tags

Not supported.

source

A GedcomSource() object describing the software that has generated the GEDCOM.`

destination

An identifier for the system expected to receive this GEDCOM.

creation_date

The creation date of the file given either as a formatted GEDCOM string, or a DateExact() object.

creation_time

The creation time of the file given either as a formatted GEDCOM string, or a Time() object.

subm_xref

The cross-reference identifier of the primary submitter.

gedcom_copyright

A copyright statement, as appropriate for the copyright laws applicable to this data.

default_language

The default language for the entire GEDCOM object.

default_place_form

The default form for place names in the GEDCOM object. A comma-separated string of jurisdictional titles. For example "City, County, State, Country".

notes

Associated notes. This can either be a Note() object, a list of them, or a character vector of notes.

note_xrefs

A character vector of relevant note record cross-reference identifiers.

update_change_dates

Whether to automatically update change dates when updating records. This happens when the record is pushed to the gedcom object.

add_creation_dates

Whether to automatically add creation dates when creating records. This happens when the record is pushed to the gedcom object.

records

An S7 object whose properties contain information about all records. Do not edit properties in capitals directly. See Details for more information.

Value

An S7 object representing a GEDCOM file.

Details

All information about records is contained in the @records property.

The @prefixes property is a named vector containing any alphanumeric string (up to 6 characters long) which will precede the number given to identify new records, of which there are currently 7 types:

Individual (INDI) Family (FAM) Source (SOUR) Repository (REPO) Multimedia (OBJE) Note (SNOTE) Submitter (SUBM)

This vector must be of a particular length with specific names. For example: c(SUBM = "U", INDI = "I", FAM = "F", SOUR = "S", REPO = "R", OBJE = "M", SNOTE = "N").

The order that these records appear in the vector will also dictate the order in which records will appear in the exported file.