Create an individual ordinance object

Create an individual ordinance object

Usage

Ordinance(
  ord_type = character(0),
  date = character(0),
  temple_name = character(0),
  place = character(0),
  ord_state = character(0),
  state_date = character(0),
  state_time = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  fam_xref = character(0)
)

Arguments

ord_type

TODO

date

The date given either as a formatted GEDCOM string, or a DateValue() object.

temple_name

TODO

place

see territory

ord_state

TODO

state_date

The ordinance date given either as a formatted GEDCOM string, or a DateExact object.

state_time

The ordinance time given either as a formatted GEDCOM string, or a Time object.

note_xrefs

A character vector of relevant note record cross-reference identifiers.

notes

Associated notes. This can either be a Note() object, a list of them, or a character vector of notes. If a character vector is provided then only the notes themselves can be recorded (and not associated information). This option is easier if associated information is not required.

citations

Associated sources. This can either be a SourceCitation() object, a list of them, or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves can be recorded (and not associated information). This option is easier if associated information is not required.

fam_xref

The cross-reference identifier of a family record.

Value

An S7 object representing a GEDCOM LDS_INDIVIDUAL_ORDINANCE.