Define properties in GEDCOM 7.0 specification

Define all properties used in S7 classes.

Arguments

full

A full address as it would appear on a mailing label, with lines separated by semi-colon and a space. For example: "The White House; 1600 Pennsylvania Avenue N.W.; Washington, DC 20500; United States of America"

adr1

Deprecated.

adr2

Deprecated.

adr3

Deprecated.

city

The city component of the address.

state

The state component of the address.

postal_code

The postal code component of the address.

country

The country component of the address.

text

A character string. New lines are created with \n.

language

A character string of language tags as defined in BCP 47.

languages

A character vector of language tags as defined in BCP 47.

media_type

The media type as defined in RFC 2045.

translations

A TranslationText object or a list of them. One for each alternate translation of the text.

title

The title of the multimedia record. This will supersede any title given in the record.

crop

Whether to crop the multimedia to a specific area.

top

The number of pixels to omit from the top side of the image.

left

The number of pixels to omit from the left side of the image.

height

The height in pixels of the cropped region.

width

The width in pixels of the cropped region.

where

A specific location within the information referenced. For a published work, this could include the volume of a multi-volume work and the page number or numbers. For a periodical, it could include volume, issue, and page numbers. For a newspaper, it could include a date, page number, and column number. For an unpublished source or micro‐ filmed works, this could be a film or sheet number, page number, or frame number. A census record might have an enumerating district, page number, line number, dwelling number, and family number. It is recommended that the data in this field be formatted comma-separated with label: value pairs

source_text

A verbatim copy of any description contained within the source. This can either be a TranslationText object, a list of them, or a character vector of text. If any TranslationText objects are provided, then you will be forced to define an associated language or media type.

fact_type

TODO

fact_phrase

Textual information that cannot be expressed in the fact type.

role

What role this person played in this fact.

role_phrase

Textual information that cannot be expressed in the role.

certainty

An enumerated value indicating the credibility of a piece of information, based on its supporting evidence. Some systems use this feature to rank multiple conflicting opinions for display of most likely information first. It is not intended to eliminate the receivers’ need to evaluate the evidence for themselves. "0" = unreliable/estimated data "1" = Questionable reliability of evidence "2" = Secondary evidence, data officially recorded sometime after event "3" = Direct and primary evidence used, or by dominance of the evidence

indi_phrase

Textual information that cannot be expressed in the @indi_xref.

relation_is

The nature of the association. This must be a value from val_roles(). If a value of "OTHER" is used, a @relation_phrase must be given.

relation_phrase

Textual information that cannot be expressed in the relation.

place_name

A comma-separated string of region names, ordered from smallest to largest. The specific meaning of each element is given by the @place_form, or in the @default_place_form of the gedcom object if there is no @place_form defined. Elements should be left blank if they are unknown, do not apply to the location, or are too specific for the region in question. For example "Baltimore, , Maryland, USA".

place_form

A comma-separated string of jurisdictional titles, which has the same number of elements as @place_form. For example "City, County, State, Country".

place_translations

A named character vector of translations of the place name. The vector values must follow the same form as the @place_name and the vector names must be a language value as defined by @language.

lat_long

The latitude and longitude of the place, separated by a space. The latitude coordinate is the direction North or South from the equator in degrees and fraction of degrees. The longitude coordinate is in degrees and fraction of degrees East or West of the zero or base meridian coordinate. For example: 18 degrees, 9 minutes, and 3.4 seconds North, 168 degrees, 9 minutes, and 3.4 seconds East would be formatted as "N18.150944 E168.150944".

fact_types

A character string indicating the types of events that were recorded in a particular source. Each event type is separated by a comma and space. For example, a parish register of births, deaths, and marriages would be BIRT, DEAT, MARR.

territory

The territory associated with the events covered. This can either be a Place object or a character string (a comma-separated string of region names, ordered from smallest to largest). If a character string is provided then only the region names can be recorded (and not associated information). This option is easier if associated information is not required.

pedigree

TODO

pedigree_phrase

TODO

confidence

TODO

confidence_phrase

TODO

call_numbers

Call number(s) used to file and retrieve items from the repository. This can either be a SourceCallNumber object, a list of them, or a character vector of call numbers. If a character vector is provided then only the call numbers themselves can be recorded (and not associated medium). This option is easier if associated information is not required.

call_number

The call number.

location

TODO

medium

TODO

medium_phrase

TODO

media_alt

A named vector of the media in alternative media forms, c(form = location)

prefix

The name prefix, e.g. Cmdr.

given

The given name or earned name.

nickname

A descriptive or familiar name that is used instead of, or in addition to, one’s proper name.

surname_prefix

Surname prefix or article used in a family name. For example in the name "de la Cruz", this value would be "de la".

surname

Surname or family name.

suffix

Name piece that appears after the given name and surname parts, e.g. Jr.

name_pieces

A PersonalNamePieces object defining the pieces of the full name.

pers_name

TODO

name_type

TODO

type_phrase

TODO

name_translations

TODO

event_type

TODO

fact_val

TODO

fact_desc

TODO

place

see territory

address

The address given either as a Address object or as a character string. This would be as written on a mailing label with new lines separated by \n.

phone_numbers

A character vector of phone numbers.

emails

A character vector of email addresses.

faxes

A character vector of fax numbers.

web_pages

A character vector of web page URLs.

agency

The organization, institution, corporation, person, or other entity that has responsibility for the associated fact. Examples are an employer of a person of an associated occupation, or an educational establishment that has awarded a scholastic award.

relig_affil

A religious denomination associated with the fact.

cause

Used in special cases to record the reasons which precipitated the fact (e.g. cause of death).

confidential

A logical value indicating whether the associated record/fact should be treated as confidential. This allows them to be excluded on export.

locked

A logical value indicating whether the associated record/fact should be treated as read-only.

private

A logical value indicating whether the associated record/fact should be treated as private. This allows them to be excluded on export.

associations

Associated individuals. This can either be a Association object or a list of them.

age

A character string that indicates the age in years, months, weeks and/or days that the individual was at the time of the fact. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 1w 3d". Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age in the corresponding phrase parameter.

husb_age

A character string that indicates the age in years, months, weeks and/or days that the husband was at the time of the fact. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 1w 3d". Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age in the corresponding phrase parameter.

wife_age

A character string that indicates the age in years, months, weeks and/or days that the wife was at the time of the fact. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 1w 3d". Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age in the corresponding phrase parameter.

age_phrase

Free text information that cannot be expressed in the individual's age.

husb_age_phrase

Free text information that cannot be expressed in the husband's age.

wife_age_phrase

Free text information that cannot be expressed in the wife's age.

adop_parent

TODO

adop_parent_phrase

TODO

pers_names

TODO

sex

The sex of the individual. Either "M" (male), "F" (female), "X" (other), or "U" (undetermined, the default),

fam_links_chil

TODO

fam_links_spou

TODO

files

A MediaFile object or a list of them. This refers to 1 or more external digital files. Grouped files should each pertain to the same context.

repo_name

The name of the repository.

facts_recorded

The facts recorded by the source. This can either be a FactsRecorded object, a list of them, or a character vector of comma-delimited fact types. If a character vector is provided then only the fact types themselves can be recorded (and not associated information). This option is easier if associated information is not required. For example, a parish register of births, deaths, and marriages would be "BIRT, DEAT, MARR". The val_fact_types() function gives a list of possible fact types.

originator

The person, agency, or entity who created the record. For a published work, this could be the author, compiler, transcriber, abstractor, or editor. For an unpublished source, this may be an individual, a government agency, church organization, or private organization.

full_title

The full title of the source.

short_title

A shortened name of the source used for sorting, filing, and retrieving records.

publication_facts

When and where the record was created. For published works, this includes information such as the city of publication, name of the publisher, and year of publication.

repo_citations

Associated repositories. This can either be a RepositoryCitation object, a list of them, or a character vector of XREFs of repository records. If a character vector is provided then only the XREFs themselves can be recorded (and not associated information). This option is easier if associated information is not required.

subm_name

The name of the submitter.

facts

TODO

non_events

TODO

spouse_sealings

TODO

ordinances

TODO

ord_type

TODO

temple_name

TODO

ord_state

TODO

xref

The cross-reference identifier for this record. You should not edit this at all as maintenance of these is done automatically.

indi_xref

The cross-reference identifier of an individual record. If the individual does not have a record, then this can be left blank and the value "@VOID@" will be used. However, you should define an @indi_phrase.

fam_xref

The cross-reference identifier of a family record.

sour_xref

The cross-reference identifier of a source record. If the source does not have a record, then this can be left blank and the value "@VOID@" will be used. However, you should describe the source in @where.

repo_xref

The cross-reference identifier of a repository record. If the repository does not have a record, then this can be left blank and the value "@VOID@" will be used. However, you should describe the repository in @notes.

media_xref

The cross-reference identifier of a multimedia record.

note_xrefs

A character vector of relevant note record cross-reference identifiers.

data_note_xrefs

A character vector of note record cross-reference identifiers relevant to the source data.

subm_xrefs

A character vector of relevant submitter record cross-reference identifiers.

husb_xref

The cross-reference identifier of the husband's individual record. If the individual does not have a record, then the value "@VOID@" can be used. However, you will need to describe the individual by using a named vector (a description can be used in either case), e.g. c("Joe Bloggs" = "@VOID@") or c("Joe Bloggs" = "@I1@")

wife_xref

The cross-reference identifier of the wife's individual record. If the individual does not have a record, then the value "@VOID@" can be used. However, you will need to describe the individual by using a named vector (a description can be used in either case), e.g. c("Joe Bloggs" = "@VOID@") or c("Joe Bloggs" = "@I1@")

chil_xrefs

The cross-reference identifier(s) of the children's individual records. If the individual does not have a record, then the value "@VOID@" can be used. However, you will need to describe the individual by using a named vector (a description can be used in either case), e.g. c("Joe Bloggs" = "@VOID@") or c("Joe Bloggs" = "@I1@")

alia_xrefs

A named character vector of relevant individual record cross-reference identifiers whose records also represent this individual. The vector names may provide a description of these records.

anci_xrefs

A character vector of relevant submitter record cross-reference identifiers who are interested in the ancestors of this individual.

desi_xrefs

A character vector of relevant submitter record cross-reference identifiers who are interested in the descendants of this individual.

unique_ids

A character vector of enduring and globally-unique identifiers. These need to be formatted in line with RFC 4122 and can be generated with uuid::UUIDgenerate().

user_ids

A character vector of user-generated identifiers. The type of the identifiers can be given in the vector names, e.g. c("Driving license number" = "ABC123")

ext_ids

A named character vector of identifiers maintained by an external authority. The names must be given as a URI. See the GEDCOM specification for more information.

media_links

Associated multimedia. This can either be a MediaLink object, a list of them, or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves can be recorded (and not associated information). This option is easier if associated information is not required.

notes

Associated notes. This can either be a Note object, a list of them, or a character vector of notes. If a character vector is provided then only the notes themselves can be recorded (and not associated information). This option is easier if associated information is not required.

data_notes

Associated notes about the source data. This can either be a Note object, a list of them, or a character vector of notes. If a character vector is provided then only the notes themselves can be recorded (and not associated information). This option is easier if associated information is not required.

citations

Associated sources. This can either be a SourceCitation object, a list of them, or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves can be recorded (and not associated information). This option is easier if associated information is not required.

hour

The hour of the day given as an integer between 0 and 23.

minute

The minute of the hour given as an integer between 0 and 59.

second

The second of the minute given as an integer between 0 and 59.

fraction

The fraction of the second given as an integer.

utc

Whether the time is in Coordinated Universal Time (UTC) (TRUE, the default) or is in local time (FALSE).

year

The year given as an integer (greater than 0).

month

The month of the year given as an integer between 1 and 12.

day

The day of the month given as an integer between 1 and 31.

bce

Whether the date is Before the Common Era. This is FALSE by default, but if TRUE, only the year should be given.

about

Whether the date is near to the date given.

calc

Whether the date is calculated from other values.

est

Whether the date is near to the date given, and is calculated from other values.

date_greg

A Gregorian date given either as a formatted GEDCOM string, or a DateGregorian object.

date_exact

An exact date given either as a formatted GEDCOM string, or a DateExact object. If not given, it will default to today's date.

date_period

A date period given either as a formatted GEDCOM string, or a DatePeriod object.

date

The date given either as a formatted GEDCOM string, or a DateValue object.

date_sort

The date given either as a formatted GEDCOM string, or a DateSorting object.

date_phrase

Textual information that cannot be expressed in the date.

start_date

The start of the period/range given either as a formatted GEDCOM string, or a DateGregorian object.

end_date

The end of the period/range given either as a formatted GEDCOM string, or a DateGregorian object.

time

The time given either as a formatted GEDCOM string, or a Time object.

created

A CreationDate object containing the date the record was created. Creating an object with no parameters sets the date to today.

updated

A ChangeDate object containing the date the record was updated. Creating an object with no parameters sets the date to today.

state_date

The ordinance date given either as a formatted GEDCOM string, or a DateExact object.

state_time

The ordinance time given either as a formatted GEDCOM string, or a Time object.

Details

This empty function serves as a single location where all properties are defined, mainly for efficiency and maintainability.