Create an individual record object

Create an individual record object

Usage

IndividualRecord(
  xref = "@ORPHAN@",
  confidential = FALSE,
  locked = FALSE,
  private = FALSE,
  user_ids = character(0),
  unique_ids = character(0),
  ext_ids = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  media_links = list(),
  created = NULL,
  updated = NULL,
  pers_names = list(),
  sex = "U",
  facts = list(),
  non_events = list(),
  ordinances = list(),
  fam_links_chil = list(),
  fam_links_spou = list(),
  subm_xrefs = character(0),
  associations = list(),
  alia_xrefs = character(0),
  anci_xrefs = character(0),
  desi_xrefs = character(0)
)

Arguments

xref

The cross-reference identifier for this record. You should not edit this at all as maintenance of these is done automatically.

confidential

A logical value indicating whether the associated record/fact should be treated as confidential. This allows them to be excluded on export.

locked

A logical value indicating whether the associated record/fact should be treated as read-only.

private

A logical value indicating whether the associated record/fact should be treated as private. This allows them to be excluded on export.

user_ids

A character vector of user-generated identifiers. The type of the identifiers can be given in the vector names, e.g. c("Driving license number" = "ABC123")

unique_ids

A character vector of enduring and globally-unique identifiers. These need to be formatted in line with RFC 4122 and can be generated with uuid::UUIDgenerate().

ext_ids

A named character vector of identifiers maintained by an external authority. The names must be given as a URI. See the GEDCOM specification for more information.

note_xrefs

A character vector of relevant note record cross-reference identifiers.

notes

Associated notes. This can either be a Note object, a list of them, or a character vector of notes. If a character vector is provided then only the notes themselves can be recorded (and not associated information). This option is easier if associated information is not required.

citations

Associated sources. This can either be a SourceCitation object, a list of them, or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves can be recorded (and not associated information). This option is easier if associated information is not required.

media_links

Associated multimedia. This can either be a MediaLink object, a list of them, or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves can be recorded (and not associated information). This option is easier if associated information is not required.

created

A CreationDate object containing the date the record was created. Creating an object with no parameters sets the date to today.

updated

A ChangeDate object containing the date the record was updated. Creating an object with no parameters sets the date to today.

pers_names

TODO

sex

The sex of the individual. Either "M" (male), "F" (female), "X" (other), or "U" (undetermined, the default),

facts

TODO

non_events

TODO

ordinances

TODO

fam_links_chil

TODO

fam_links_spou

TODO

subm_xrefs

A character vector of relevant submitter record cross-reference identifiers.

associations

Associated individuals. This can either be a Association object or a list of them.

alia_xrefs

A named character vector of relevant individual record cross-reference identifiers whose records also represent this individual. The vector names may provide a description of these records.

anci_xrefs

A character vector of relevant submitter record cross-reference identifiers who are interested in the ancestors of this individual.

desi_xrefs

A character vector of relevant submitter record cross-reference identifiers who are interested in the descendants of this individual.

Value

An S7 object representing a GEDCOM INDIVIDUAL_RECORD.